task(test-scenario): re org, split
